load XML

+++ include/pt/mesh.h
@@ -17,12 +17,12 @@ public:
         const std::vector<TriVertex>& vertices,
         const std::vector<TriNormal>& normals,
         const std::vector<TriUV>& uvs,
-        const std::vector<int>& mat_ids
+        const std::vector<uint32_t>& mat_ids
     ) : m_name(name), m_tri_vertices(vertices), m_tri_normals(normals), m_tri_uvs(uvs), m_mat_ids(mat_ids){ }
 
     std::string getName() { return m_name; }
 
-    int getTriangleCount() { return m_tri_vertices.size(); }
+    size_t getTriangleCount() { return m_tri_vertices.size(); }
 
 private:
     std::string m_name;
@@ -30,7 +30,7 @@ private:
     std::vector<TriVertex> m_tri_vertices;
     std::vector<TriNormal> m_tri_normals;
     std::vector<TriUV> m_tri_uvs;
-    std::vector<int> m_mat_ids;
+    std::vector<uint32_t> m_mat_ids;
 };
 
 }
